+++ st2api/st2api/controllers/v1/actionalias.py
@@ -156,12 +156,15 @@ class ActionAliasController(resource.ContentPackResourceController):
 
         return action_alias_api
 
-    @request_user_has_resource_db_permission(permission_type=PermissionType.ACTION_MODIFY)
+    @request_user_has_resource_db_permission(permission_type=PermissionType.ACTION_ALIAS_MODIFY)
     @jsexpose(arg_types=[str], body_cls=ActionAliasAPI)
-    def put(self, action_alias_ref_or_id, action_alias):
+    def put(self, action_alias, action_alias_ref_or_id):
         action_alias_db = self._get_by_ref_or_id(ref_or_id=action_alias_ref_or_id)
         LOG.debug('PUT /actionalias/ lookup with id=%s found object: %s', action_alias_ref_or_id,
                   action_alias_db)
+
+        if not hasattr(action_alias, 'id'):
+            action_alias.id = None
 
         try:
             if action_alias.id is not None and action_alias.id is not '' and \