+++ CHANGELOG.rst
@@ -56,6 +56,7 @@ in development
 * Add support for `st2 login` and `st2 whoami` commands. These add some additional functionality
   beyond the existing `st2 auth` command and actually works with the local configuration so that
   users do not have to.
+* Fix action alias update API endpoint. (bug fix)
 
 2.1.1 - December 16, 2016
 -------------------------